requirements.txt
python-telegram-bot==20.0
yfinance==0.2.20